XWIKI-22786: Update or mark method User#isUserInGroup as deprecated
Deprecate com.xpn.xwiki.api.User#isUserInGroup(String) and
com.xpn.xwiki.user.api.XWikiUser#isUserInGroup(String, XWikiContext) and move them to
xwiki-platform-legacy-oldcore, which re-adds them by weaving xwiki-platform-oldcore with
inter-type declarations. Both aspects are privileged since the methods use private members of the
classes they are declared on.

The recommended replacement is the org.xwiki.user.group.GroupManager component: unlike these
methods it takes a group reference (and is thus not restricted to the current wiki) and it can
optionally take sub-groups into account.

The two remaining callers keep the current semantics (direct members only, so no behavior
changes):
* getusersandgroups.vm now uses the user.group script service. The current user reference is
  checked explicitly since it is null for the guest user.
* GroupMessageStreamNotificationFilter is left untouched: since it is itself legacy code, its
  module now depends on xwiki-platform-legacy-oldcore rather than xwiki-platform-oldcore so that
  it still sees the deprecated method.

xwiki-platform-oldcore itself skips Revapi because xwiki-platform-legacy-oldcore wraps it and runs
the check on the woven jar, which still exposes both methods. Modules depending on it are checked
against the unwoven jar though, and xwiki-platform-legacy-extension-handler-xar is the one module
that exposes both classes in its API and therefore reported the two methods as removed. It now
depends on xwiki-platform-legacy-oldcore too, so that its check also compares against the woven
jar. That comparison is asymmetric until the next release, since the version it is compared
against still depends on the unwoven jar, hence the two ignores for the inter-type declarations
that only the woven jar has.
